src/backend/pyproject.toml
[project]
name = "project-transformers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.11",
    "torch>=2.8.0",
    "transformers>=4.56.0",
]
